+++ Web_app/MUgituApi/src/main/java/Mooving/MUgituApi/dao/tipoAveria/TipoAveriaRepository.java
@@ -3,5 +3,5 @@ package Mooving.MUgituApi.dao.tipoAveria;
 import Mooving.MUgituApi.entities.TipoAveria;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TipoAveriaRepository extends JpaRepository<TipoAveria, Long> {
+public interface TipoAveriaRepository extends JpaRepository<TipoAveria, Integer> {
 }
